Ringbuffer: Move files into a package
